chore: stop tracking build artifacts
