+++ docs/CODE_REVIEW.md
@@ -1,0 +1,25 @@
+# Code Review: Simple Air Comfort Card
+
+## Summary
+The card implementation is impressively well-documented and thoughtfully structured. I especially appreciated the thorough configuration normalization in `setConfig`, the defensive handling of legacy YAML keys, and the responsive typography scaling driven by the `ResizeObserver` hook in `connectedCallback`.
+
+## Strengths
+- **Robust configuration parsing.** Flattening grouped YAML keys and normalizing display-unit tokens makes the runtime config resilient to editor output and manual YAML edits.【F:src/card/simple-air-comfort-card.ts†L174-L260】
+- **Responsive layout strategy.** Using a `ResizeObserver` to derive `--sac-scale` keeps the dial legible at a wide range of card widths without relying on viewport queries.【F:src/card/simple-air-comfort-card.ts†L118-L143】
+- **Clear separation of physics vs. view-model.** The render method stages data, labels, gradients, and geometry before passing a compact object to `#face`, which keeps the template manageable.【F:src/card/simple-air-comfort-card.ts†L309-L520】
+
+## Findings
+### 1. Wind chill outside its validity range
+The NWS/Environment Canada wind-chill equation is only validated for air temperatures ≤ 10 °C and wind speeds ≥ 4.8 km/h. As written, the helper returns the regression result even when conditions fall outside that domain, which can produce noticeably colder “feels like” values than the ambient temperature during mild or calm conditions.【F:src/card/simple-air-comfort-card.ts†L614-L619】 Consider short-circuiting to the ambient temperature when the inputs are outside the published limits so the UI doesn’t flag false cold alarms.
+
+### 2. Heat index without low-range adjustments
+Similarly, the Rothfusz heat-index regression is documented for T ≥ 80 °F and RH ≥ 40 %. Below that envelope the U.S. NWS applies additional adjustments (or simply returns the air temperature). Omitting those adjustments leads to a few degrees of apparent-temperature inflation on pleasant days when users switch to the heat-index mode.【F:src/card/simple-air-comfort-card.ts†L621-L632】 Adding the standard correction tables (or clamping to the dry-bulb temperature when below-threshold) would align the output more closely with expectations.
+
+### 3. Placeholder value formatting
+When temperature-derived numbers are unavailable, `#formatNumber` returns an em dash, but the string concatenation still appends the unit suffix, yielding outputs like “— °C” in the card’s corner labels.【F:src/card/simple-air-comfort-card.ts†L464-L466】 You might want to suppress the unit when the numeric portion is missing so the placeholders read simply “—”.
+
+## Suggestions
+- Memoize `#bandThresholds()` across a render cycle if you notice performance pressure; it re-derives the same ladder multiple times when both fallback and main branches execute in a single update.【F:src/card/simple-air-comfort-card.ts†L396-L520】
+- In the editor, consider showing both the selected preset and the currently applied cap range next to each temperature anchor. The existing helper text explains it, but visualizing the cap window could help new users grok the coupling faster.【F:src/editor/simple-air-comfort-card-editor.ts†L298-L360】
+
+Overall this is a polished card—these tweaks would mainly tighten the edge-case behavior for the derived comfort metrics.
