+++ docs/typescript-advantages.md
@@ -1,0 +1,11 @@
+# TypeScript Advantages for Simple Air Comfort Card
+
+While the card currently ships as JavaScript, adopting TypeScript could unlock several benefits if you are willing to introduce a compile step:
+
+- **Static typing reduces runtime surprises** – Type annotations catch incompatible values during development, safeguarding interactions with Home Assistant APIs and custom element properties before they ship.
+- **Improved editor tooling** – Editors like VS Code provide richer IntelliSense, auto-completion, and refactor support when type metadata is available.
+- **Easier refactors** – Type-aware renames and interface definitions give confidence when splitting logic across modules or updating configuration schemas.
+- **Safer integration with external libraries** – Declaration files document expectations and highlight mismatches in third-party code, prompting fixes early.
+- **Gradual adoption is possible** – TypeScript lets you opt files in one-by-one, so you can start with high-churn modules (such as card logic) without converting the entire repo at once.
+
+Trade-offs include introducing a build step (e.g., `tsc` or Babel), maintaining type definitions, and keeping generated `.js` artifacts aligned with HACS distribution requirements. If those costs are acceptable, TypeScript can make ongoing maintenance smoother and less error-prone.
